pyproject.toml
[project]
name = "gemini-voice-agent"
version = "0.1.0"
description = "Voice agent powered by Gemini Live native audio — no STT/TTS needed"
requires-python = ">=3.10"
dependencies = [
    "pipecat-ai[webrtc,websocket,silero,google,runner]>=0.0.101",
    "pipecatcloud>=0.2.18",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "aiohttp>=3.9.0",
    "uvicorn>=0.27.0",
    "fastapi>=0.109.0",
    "google-api-python-client>=2.100.0",
    "google-auth>=2.25.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
]

[dependency-groups]
dev = [
    "ruff~=0.12.1",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["I"]
